Add first basic end to end test
