+++ test/end_to_end/steps/test_parent_steps.py
@@ -1,0 +1,13 @@
+from time import sleep
+
+from pytest_bdd import given, scenarios
+
+from ..utils.pages.homepage import HomePage
+
+scenarios("../features/smoke.feature")
+
+
+@given("I an authorised user")
+def login():
+    """Login to the application."""
+    HomePage().navigate_to_next_page()
