--- test/end_to_end/utils/navigate.py
@@ -1,2 +1,0 @@
-def navigate_to_page():
-    pass
