+++ test/end_to_end/utils/pages/page.py
@@ -1,7 +1,19 @@
+from selenium.webdriver.common.by import By
+
 from ..drivers.chrome_driver import get_driver
 
 
-class page:
+class Page:
     """Base class to initialize the base webpage that will be called from all pages"""
 
     url_subdirectory: str
+    page_number: int
+
+    def navigate_to_next_page(self):
+        """Navigate to the next page"""
+        click_next_button()
+
+
+def click_next_button():
+    """Click the next button"""
+    return get_driver().find_element(By.CLASS_NAME, "nhsuk-button").click()
